refactor(rooms): rename Agora identifiers to live-rooms; @syra.fm/live 0.2.0
Agora is gone — everything is Syra. Mention frontend rename (behavior identical):
lib/agoraConfig.tsx -> lib/liveConfig.tsx; agoraConfig -> liveConfig;
AgoraProvider -> LiveConfigProvider; useAgoraConfig -> useLiveConfig;
createAgoraService -> createRoomsService; Agora/AgoraActive icons ->
LiveRoomsIcon/LiveRoomsIconActive. Bump @syra.fm/live ^0.2.0. i18n agora.* keys
left (shared lookup contract with the engine fallback).
